requirements.txt
streamlit
opencv-python
ultralytics
pymongo
python-dotenv
requests
bcrypt
numpy
pillow